feat(형진):출석기능 구현 중

+++ backend/ASAF/src/main/java/com/ASAF/dto/AttendanceDTO.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 @AllArgsConstructor
 @ToString
 public class AttendanceDTO {
-    private Long attendanceId;
+    private int attendanceId;
     private MemberDTO memberDTO;
     private String attended;
     private String entryTime;
